refactor(app): refactor questions filtering
