+++ apps/app/src/lib/order.ts
@@ -1,12 +1,20 @@
 import { QueryParam } from "../types";
 
-const ordersBy = ["acceptedAt", "level", "votesCount"] as const;
-const orders = ["asc", "desc"] as const;
-export const levels = ["junior", "mid", "senior"] as const;
-
-export type Level = typeof levels[number];
-
+export const ordersBy = ["acceptedAt", "level", "votesCount"] as const;
+export const orders = ["asc", "desc"] as const;
 export const DEFAULT_SORT_BY_QUERY = "acceptedAt*desc";
+
+export const sortByLabels: Record<`${OrderBy}*${Order}`, string> = {
+	"acceptedAt*asc": "od najstarszych",
+	"acceptedAt*desc": "od najnowszych",
+	"level*asc": "od najprostszych",
+	"level*desc": "od najtrudniejszych",
+	"votesCount*asc": "od najmniej popularnych",
+	"votesCount*desc": "od najpopularniejszych",
+};
+
+type OrderBy = typeof ordersBy[number];
+type Order = typeof orders[number];
 
 export const parseQuerySortBy = (query: QueryParam) => {
 	if (typeof query !== "string") {
@@ -20,18 +28,4 @@ export const parseQuerySortBy = (query: QueryParam) => {
 	}
 
 	return { orderBy, order };
-};
-
-export const parseQueryLevels = (query?: QueryParam | null) => {
-	if (typeof query !== "string") {
-		return null;
-	}
-
-	const splittedQuery = query.split(",");
-
-	if (!splittedQuery.every((level) => levels.includes(level))) {
-		return null;
-	}
-
-	return splittedQuery as Level[];
 };
